move tests from basic to py_config
