+++ pyscriptjs/tests/integration/test_01_basic.py
@@ -73,76 +73,6 @@ class TestBasic(PyScriptTest):
         )
         assert self.console.log.lines == [self.PY_COMPLETE, "true false", "<div></div>"]
 
-    def test_paths(self):
-        self.writefile("a.py", "x = 'hello from A'")
-        self.writefile("b.py", "x = 'hello from B'")
-        self.pyscript_run(
-            """
-            <py-config>
-                [[fetch]]
-                files = ["./a.py", "./b.py"]
-            </py-config>
-
-            <py-script>
-                import js
-                import a, b
-                js.console.log(a.x)
-                js.console.log(b.x)
-            </py-script>
-            """
-        )
-        assert self.console.log.lines == [
-            self.PY_COMPLETE,
-            "hello from A",
-            "hello from B",
-        ]
-
-    def test_paths_that_do_not_exist(self):
-        self.pyscript_run(
-            """
-            <py-config>
-                [[fetch]]
-                files = ["./f.py"]
-            </py-config>
-            """
-        )
-        assert self.console.error.lines == ["Failed to load resource: net::ERR_FAILED"]
-        assert self.console.warning.lines == [
-            "Caught an error in fetchPaths:\r\n TypeError: Failed to fetch"
-        ]
-
-        errorContent = """PyScript: Access to local files
-        (using "Paths:" in &lt;py-config&gt;)
-        is not available when directly opening a HTML file;
-        you must use a webserver to serve the additional files."""
-
-        inner_html = self.page.locator(".py-error").inner_html()
-        assert errorContent in inner_html
-
-    def test_paths_from_packages(self):
-        self.writefile("utils/__init__.py", "")
-        self.writefile("utils/a.py", "x = 'hello from A'")
-        self.pyscript_run(
-            """
-            <py-config>
-                [[fetch]]
-                from = "utils"
-                to_folder = "pkg"
-                files = ["__init__.py", "a.py"]
-            </py-config>
-
-            <py-script>
-                import js
-                from pkg.a import x
-                js.console.log(x)
-            </py-script>
-            """
-        )
-        assert self.console.log.lines == [
-            self.PY_COMPLETE,
-            "hello from A",
-        ]
-
     def test_packages(self):
         self.pyscript_run(
             """
